Add files via upload
This is basic regression analysis to identify the impact of discount, featuring the product, and price on the weekly sales of juices in a chain of grocery stores. I first run the regression with each variable separately. I then run the regression with multiple variables and, finally, run the regression with interaction.